refactoring global graph
