+++ src/fgsp/graph/global_graph.py
@@ -1,6 +1,5 @@
 #! /usr/bin/env python3
 import time
-import multiprocessing
 
 import numpy as np
 from scipy import spatial
@@ -17,59 +16,6 @@ from src.fgsp.common.visualizer import Visualizer
 from src.fgsp.graph.base_graph import BaseGraph
 from src.fgsp.common.utils import Utils
 from src.fgsp.common.visualizer import Visualizer
-
-
-def process_poses(poses, tree, w_func, i):
-    if len(poses) == 0:
-        return
-    max_pos_dist = 6.0
-    nn_indices = tree.query_ball_point(
-        poses[i, 0:3], r=max_pos_dist, p=2)
-
-    return nn_indices, [w_func(poses[i, 0:3], poses[nn_i, 0:3]) for nn_i in nn_indices]
-
-
-def compute_distance_weights(coords_lhs, coords_rhs):
-    sigma = 1.0
-    normalization = 2.0*(sigma**2)
-    dist = spatial.distance.euclidean(coords_lhs, coords_rhs)
-    # print(f'Distance between {coords_lhs} and {coords_rhs} is {dist}')
-
-    return np.exp(-dist/normalization)
-
-
-def compute_so3_weights(pose_lhs, pose_rhs):
-    R_lhs = Utils.convert_quat_to_rotation(pose_lhs[3:7])
-    R_rhs = Utils.convert_quat_to_rotation(pose_rhs[3:7])
-    rot_diff = np.matmul(R_lhs, R_rhs.transpose())
-    return np.trace(rot_diff)
-
-
-def compute_se3_weights(poses_lhs, poses_rhs):
-    T_G_lhs = Utils.convert_pos_quat_to_transformation(
-        poses_lhs[0:3], poses_lhs[3:7])
-    T_G_rhs = Utils.convert_pos_quat_to_transformation(
-        poses_rhs[0:3], poses_rhs[3:7])
-
-    pose1 = SE3.from_matrix(T_G_lhs)
-    pose2 = SE3.from_matrix(T_G_rhs)
-
-    Xi_12 = (pose1.inv().dot(pose2)).log()
-    W = np.eye(4, 4)
-    W[0, 0] = 10
-    W[1, 1] = 10
-    W[2, 2] = 1
-    W[3, 3] = 3
-    inner = np.trace(
-        np.matmul(np.matmul(SE3.wedge(Xi_12), W), SE3.wedge(Xi_12).transpose()))
-
-    # Equal weighting for rotation and translation.
-    # inner = np.matmul(Xi_12.transpose(),Xi_12)
-
-    dist = np.sqrt(inner)
-    sigma = 1.0
-    normalization = 2.0*(sigma**2)
-    return np.exp(-dist/normalization)
 
 
 class GlobalGraph(BaseGraph):
@@ -235,32 +181,6 @@ class GlobalGraph(BaseGraph):
     def get_graph(self):
         return self.G
 
-    def create_adjacency_from_poses(self, poses):
-        n_coords = poses.shape[0]
-        adj = np.zeros((n_coords, n_coords))
-        tree = spatial.KDTree(poses[:, 0:3])
-
-        indices = np.arange(0, n_coords)
-
-        if self.config.use_se3_computation:
-            func = partial(process_poses, poses, tree, compute_se3_weights)
-        elif self.config.use_so3_computation:
-            func = partial(process_poses, poses, tree, compute_so3_weights)
-        else:
-            func = partial(process_poses, poses, tree,
-                           compute_distance_weights)
-
-        n_cores = multiprocessing.cpu_count()
-        with Pool(n_cores) as p:
-            for idx, (nn_indices, weights) in zip(indices, p.map(func, indices)):
-                for nn_i, w in zip(nn_indices, weights):
-                    if nn_i != idx:
-                        adj[idx, nn_i] = w
-
-        adj[adj < 0] = 0
-        assert np.all(adj >= 0)
-        return adj
-
     def compute_temporal_decay(self, timestmap_lhs, timestamp_rhs):
         ts_diff_s = Utils.ts_ns_to_seconds(
             np.absolute(timestmap_lhs - timestamp_rhs))
@@ -308,43 +228,6 @@ class GlobalGraph(BaseGraph):
 
         assert np.all(self.adj >= 0)
         self.G.compute_fourier_basis()
-
-    def reduce_every_other(self):
-        n_nodes = np.shape(self.coords)[0]
-        return np.arange(0, n_nodes, 2)
-
-    def reduce_largest_ev_positive(self, take_n):
-        idx = np.argmax(np.abs(self.G.U))
-        idx_vertex, idx_fourier = np.unravel_index(idx, self.G.U.shape)
-        indices = []
-        for i in range(0, take_n):
-            if (self.G.U[i, idx_fourier] >= 0):
-                indices.append(i)
-        return indices
-
-    def reduce_largest_ev_negative(self, take_n):
-        idx = np.argmax(np.abs(self.G.U))
-        idx_vertex, idx_fourier = np.unravel_index(idx, self.G.U.shape)
-        indices = []
-        for i in range(0, take_n):
-            if (self.G.U[i, idx_fourier] < 0):
-                indices.append(i)
-        return indices
-
-    def reduce_largest_ev(self, take_n):
-        Logger.LogInfo(f'GlobalGraph: Reducing to largest {take_n} EVs.')
-        indices = []
-        ev = np.abs(self.G.U)
-        for i in range(0, take_n):
-            idx = np.argmax(ev)
-            idx_vertex, idx_fourier = np.unravel_index(idx, self.G.U.shape)
-            if ev[idx_vertex, idx_fourier] == -1:
-                Logger.LogWarn(
-                    f'GlobalGraph: Warning Could not reduce to requested number of nodes: {len(indices)}/{take_n}.')
-                return indices
-            ev[idx_vertex, :] = -1
-            indices.append(idx_vertex)
-        return indices
 
     def to_graph_msg(self):
         graph_msg = Graph()
